We should not reflect changes made to ignored filters.
Prior to this changeset, changes made to any filter were immediately
reflected in the state of the Local Node Tree.

This behavior has been changed so that filter changes are deferred while
a given node is ignored. These deferred changes are executed later when
the node, by whatever means, becomes unignored.